front/requirements.txt
streamlit
streamlit-option-menu
pandas
polars
pyarrow
plotly
requests
python-dotenv